Introduce /month and /day redirect pages
Also turn About page into new index
