+++ pages/day/index.js
@@ -1,0 +1,9 @@
+import { getCurrentDayRedirect } from "./[day]";
+
+export default function IndexPage() {
+  return null;
+}
+
+export async function getServerSideProps() {
+  return getCurrentDayRedirect();
+}
