+++ pages/month/index.js
@@ -1,0 +1,9 @@
+import { getCurrentMonthRedirect } from "./[month]";
+
+export default function IndexPage() {
+  return null;
+}
+
+export async function getServerSideProps() {
+  return getCurrentMonthRedirect();
+}
